Add new email and postal address for centralisation
Ticket: https://trello.com/c/ODIpiZet

Follow-up to PR #1152 to prepare the centralisation.

A bit of refactor to better organise the email and postal address methods into a separate concern that will handle the logic to know if the court is centralised or not, and return the details accordingly.

This is still not in use, as the new method `#documents_email` it is not used yet and the `#full_address` keeps returning the same address as before.

In a follow-up PR we can start enabling this.
