+++ lib/tasks/mutant.rake
@@ -88,7 +88,6 @@ def models
     Applicant
     Solicitor
     User
-    Court
     PaymentIntent
     CompletedApplicationsAudit
   ).freeze
