+++ spec/models/court_spec.rb
@@ -8,36 +8,14 @@ describe Court do
       "name" => 'Court name',
       "slug" => 'court-slug',
       "email" => 'family@court',
-      "address" => address,
+      "address" => 'address',
       "cci_code" => 123,
       "gbs" => 'X123',
     }
   }
 
-  let(:address) {
-    {
-      "address_lines" => ["351 Silbury Boulevard", "Witan Gate East"],
-      "town" => "Central Milton Keynes",
-      "postcode" => "MK9 2DT",
-    }
-  }
-
   describe 'REFRESH_DATA_AFTER' do
     it { expect(described_class::REFRESH_DATA_AFTER).to eq(72.hours) }
-  end
-
-  describe 'Centralised courts (smoke test)' do
-    it 'returns the list of slugs taking part in the centralisation' do
-      expect(
-        subject.send(:centralised_slugs)
-      ).to match_array(%w(
-        brighton-county-court
-        chelmsford-county-and-family-court
-        leeds-combined-court-centre
-        medway-county-court-and-family-court
-        west-london-family-court
-      ))
-    end
   end
 
   describe '.build' do
@@ -50,7 +28,7 @@ describe Court do
     end
 
     it 'sets the address' do
-      expect(subject.address).to eq(address)
+      expect(subject.address).to eq('address')
     end
 
     it 'sets the cci_code' do
@@ -311,31 +289,6 @@ describe Court do
 
           expect(subject).to eq(court)
         end
-      end
-    end
-  end
-
-  describe '#full_address' do
-    let(:address) { { 'address_lines' => address_lines, 'town' => 'town', 'postcode' => 'postcode' } }
-    let(:address_lines){ ['line 1', 'line 2'] }
-
-    it 'returns a flattened array of name, address_lines, town and postcode' do
-      expect(subject.full_address).to eq(['Court name', 'line 1', 'line 2', 'town', 'postcode'])
-    end
-
-    context 'when any lines are duplicated' do
-      let(:address_lines){ ['Court name', 'postcode'] }
-
-      it 'removes the duplicates' do
-        expect(subject.full_address).to eq(['Court name', 'postcode', 'town'])
-      end
-    end
-
-    context 'when any lines are blank' do
-      let(:address_lines){ ['line 1', ''] }
-
-      it 'removes the blank lines' do
-        expect(subject.full_address).to eq(['Court name', 'line 1', 'town', 'postcode'])
       end
     end
   end
@@ -601,28 +554,6 @@ describe Court do
       it 'returns false' do
         expect(subject).to receive(:gbs).and_return('unknown')
         expect(subject.gbs_known?).to eq(false)
-      end
-    end
-  end
-
-  describe '#centralised?' do
-    before do
-      allow(subject).to receive(:slug).and_return(slug)
-    end
-
-    context 'for a centralised court' do
-      let(:slug) { 'west-london-family-court' }
-
-      it 'returns true' do
-        expect(subject.centralised?).to eq(true)
-      end
-    end
-
-    context 'for a non-centralised court' do
-      let(:slug) { 'derby-combined-court-centre' }
-
-      it 'returns false' do
-        expect(subject.centralised?).to eq(false)
       end
     end
   end
